Run E2E tests against dev database and hosting environment
I originally introduced [1] a separate E2E hosting environment primarily
so that the Playwright end-to-end tests could run against a dedicated
database and wouldn't interfere with, or be affected by, any data that
was manually added to the application database in development.

The environment was also deliberately configured with an invalid
SendGrid API key to prevent emails from accidentally being sent by
end-to-end tests.

However, since multiple end-to-end tests are typically run in parallel,
they all have to be written to be tolerant of database records being
created and updated by concurrent tests and to clean up any database
records that they create. This means that running them against a
separate database isn't as important as I first thought.

Separately we've also since switched from SendGrid to Azure
Communication Services for sending emails and, as part of this, we no
longer have a dedicated API key for sending emails and therefore cannot
use a fake API key to prevent email sending.

Furthermore, having a separate database and hosting environment for E2E
tests added some complexity to certain development workflows, including
having to manually delete the E2E database after every schema change so
that it would get recreated on the next run.

I've therefore decided to reverse that decision and instead have the E2E
tests run against the application database and development hosting
environment. This means that we now only have one application database
to keep in sync with migrations and no longer have to run a separate
instance of the application for the end-to-end tests.

[1] 4116ba39338e7ecc5c787fc96d895ec1c24d678a