Create initial admin user when seeding app database for development
In development we're already automatically seeding the application
database with users for the end-to-end tests, so we might as well use
the same mechanism to create an admin user for local dev testing too (as
opposed to requiring developers to sign in with the E2E tests users or
manually insert a user record themselves).

The email address for this user is dev@example.com and the initial
password is 'dev-pass'.