Add Mailpit sender
This will become the default email sender in development and will also
eventually be used for automated end-to-end testing of user journeys
that involve emails.

Mailpit can receive messages via SMTP or its proprietary HTTP API. Since
I don't have any plans to use SMTP for sending emails more generally,
and because the the SMTP client built into .NET is not recommended for
use in new development, I decided that the proprietary HTTP API was the
most appropriate choice in this case.

The Mailpit API is assumed to be hosted at http://localhost:8025 by
default, but this can be changed through the Email > MailpitServer
config value.