Add security event manager with method to create a security event
We currently have the code for inserting a security event duplicated in
a few places across the application.

Moving that code into a dedicated service reduces duplication and
improves testability at the slight cost of no longer being able to
insert the event record as part of the same transaction as related
changes (e.g. the update to the user record on a password change).