Add mutation for creating a test user in end-to-end tests
Since this mutation is only intended for use in end-to-end tests, it is
excluded from the GraphQL schema when the app is not running in the
development runtime environment.

It is also protected by an admin-only authorization policy to ensure
that it doesn't pose a security risk even if it does become accidentally
exposed through misconfiguration in production.