Allow users to be created with an initial password
For security reasons we only want to set initial passwords when setting
up users for automated end-to-end tests. All other scenarios it is more
appropriate for users to set choose their own password via the password
reset flow (or, in future, via a separate user invitation flow).

The new `Password` property on `NewUserAttributes` is therefore excluded
from the corresponding GraphQL input type so that it cannot be used when
creating users via the existing `createUser` mutation.