Add user manager method to create a user for end-to-end testing
This method will be used to create dedicated users for end-to-end tests
that cannot safely use the seeded `e2e-user@example.com` and
`e2e-admin@example.com` user accounts. This will include tests that
modify the user account, such as end-to-end tests for the password
change and reset flows.

To ensure uniqueness of the name and email address, the method uses the
random token generator to create a pseudo-random suffix. In the very
unlikely event that the randomly generated suffix collides with an
existing user, a unique key constraint violation will occur on the email
column which this method automatically handles by retrying with another
randomly generated suffix.

Although the intention is not to provide any means of invoking this
method in production, there is always the risk of misconfiguration
leading to the application running using the development hosting
environment in production. To mitigate this risk, the users created by
this method have a randomly generated password that is derived
independently of the suffix. This ensure that even if a test user
created through this method ends up in a production database, there
should be no way for an attacker to use it to authenticate without
access to the password that was returned at the point of creation.