Add user manager method for hard-deleting a test user
This method will be used to delete users that have been created in
end-to-end tests. It is not intended to be used in production as it
deletes security event records that should be retained as an audit
trail.

The method does not contain any code to explicitly delete password reset
tokens as these are automatically taken care of by the cascade delete on
the foreign key constraint. The unit test still checks for this,
however, to make the intent clear and ensure that the behaviour of the
method is not inadvertently changed if the foreign key is modified in
the future.