Add mutation for deleting a test user in end-to-end tests
Since this mutation is only intended for use in end-to-end tests, it is
excluded from the GraphQL schema when the app is not running in the
development runtime environment.

It is also protected by an admin-only authorization policy to mitigate
the risk of it becoming accidentally exposed through misconfiguration of
the runtime environment in production.