Switch order of assertions in `ApiAssert.SuccessResponse`
Generally speaking if a GraphQL request fails then both of these
assertions will fail. However the failure message from the assertion
that there is no errors field is more concise and therefore usually more
convenient for quickly identifying the cause of the failure, so it makes
sense to check that assertion first.