Add workflow for publishing as a container
For now, this workflow will be manually triggered whenever a new
container image is needed.

I'm using MSBuild (as opposed to a Dockerfile) to build the container
image since that appears to be Microsoft's current recommended approach
[1].

I've specified `noble-chiseled-extra` as the container family to get the
smallest available ASP.NET Core 10 ('distroless') base image with
globalization support [2].

I'm using the `redhat-actions/push-to-registry` action to push the image
to the GitHub Container Registry as it conveniently provides the digest
of the pushed image as an output parameter, which we can then use to
create the attestation.

Although that action supports pushing images from the Docker image
store, it does so by copying the image to the Podman image store and
then pushing it up from there. To avoid that extra copy step, I've used
the `LocalRegistry` property to instruct MSBuild to save the image
directly in the Podman image store.

[1] https://devblogs.microsoft.com/dotnet/streamline-container-build-dotnet-8/
[2] https://github.com/dotnet/dotnet-docker/blob/main/documentation/sample-image-size-report.md