Use `MapStaticAssets` instead of `UseStaticFiles`
ASP.NET Core 9 introduced `MapStaticAssets` as a drop-in replacement for
`UseStaticFiles` [1] with built-in support for build time compression
and fingerprinting of static assets, as well automatically adding
optimal cache control headers.

This commit therefore sets up the app to use `MapStaticAssets` instead
of `UseStaticFiles`, although further changes will be needed across
subsequent commits to take advantage of the new features.

[1] https://learn.microsoft.com/en-us/aspnet/core/release-notes/aspnetcore-9.0?view=aspnetcore-10.0#static-asset-delivery-optimization