Use fingerprinting from MapStaticAssets for fonts and images
Since ASP.NET Core now has built-in support for static asset
fingerprinting via `MapStaticAssets`, it is no longer necessary to
fingerprint images and fonts in the Gulp build process.

This commit therefore removes the corresponding code from the Gulpfile
and updates the Razor views to use the static asset paths directly
instead of calling `IAssetHelper` as the built-in ASP.NET Core `img` and
`link` tag helpers will now take care of injecting the fingerprints into
the URLs.

The one thing that we are longer able to do as a result of this change
is injecting the fingerprinted font URLs into the CSS files. However,
this is unlikely to make much of a difference in practice as the default
cache control headers that are added through `MapStaticAssets` ensure
that the fonts are still cached by the browsers for at least an hour and
then revalidated using their fingerprint as the ETag (and therefore only
re-downloaded if they've actually been modified).