Stop generating source maps for style sheets
The `gulp-sourcemaps` package we've been relying on to generate the
source maps for stylesheets in development was last updated 5 years ago
and depends on a version of PostCSS with a known vulnerability [1].

In practice I don't think I've ever made use of these source maps, not
least because the un-minified style sheets used in development are
perfectly readable. It therefore seems more appropriate to stop
generating the source maps completely rather than spending time trying
to find an alternative solution that doesn't involve unmaintained and
vulnerable packages.

[1] https://github.com/advisories/GHSA-7fh5-64p2-3v2j