Add workflow for publishing as a ZIP archive
This workflow is intended to complement the 'Publish Container' workflow
I've just added.

It builds a framework-dependent deployment if the app and uploads a ZIP
archive of the output as a build artifact.

Like the 'Publish Container' workflow, this workflow will be triggered
manually whenever a new non-containerised deployment is required.