Publish containers for both x64 and ARM64
Up to now we've only been publishing container images for the x64
architecture of the ubuntu-24.04 workflow runner.

However, since the production instance of Buttercup is hosted on an
ARM64-based web server, in order to use the containers in production we
need to publish images for ARM64 too.

MSBuild is actually perfectly capable of building multi-architecture
container images all in one go [1], but not when pushing to a local
container store (which, as far as I'm aware, is currently the only way
to get the pushed image digests needed to produce attestations).

This commit therefore updates the Publish Containers workflow to use
MSBuild to build the architecture-specific images before using the
Podman CLI manifest commands to add them a multi-architecture image.

I originally tried using `subject-checksums` [2] to create a single
attestation covering all three images, but this resulted in an "Invalid
Argument - duplicate subject in statement" error. Given that the
documentation for `actions/attest-build-provenance` states that the
subject name for container attestations must be the fully-qualified
image name without tags [3], I believe that makes it necessary to create
a separate attestation for each image.

[1] https://learn.microsoft.com/en-us/dotnet/core/containers/publish-configuration#multi-architecture-images
[2] https://github.com/actions/attest-build-provenance?tab=readme-ov-file#identify-subjects-with-checksums-file
[2] https://github.com/actions/attest-build-provenance?tab=readme-ov-file#container-image