Defer MySQL server version detection to after application start up
Previously the application would halt with the following unhandled
exception on startup if the MySQL server indicated by the connection
string was unavailable:

    MySqlConnector.MySqlException (0x80004005): Unable to connect to
    any of the specified MySQL hosts.

This was happening because `ServerVersion.AutoDetect`, which relies on
the MySQL server being available to detect the server version, was being
called immediately during service registration.

Wrapping the `ServerVersion.AutoDetect` call in a lazy factory defers it
until a database context is first needed, ensuring that failure doesn't
trigger an unhandled exception during application startup. Setting the
lazy mode to `PublicationOnly` is also important to ensure that
exceptions are not cached, allowing for retries is the MySQL server was
temporarily unavailable on the first attempt.