Use Podman CLI to push container images
`redhat-actions/push-to-registry` felt a bit over-engineered as it runs
commands to:

- Create a temporary Podman image store for copying images from Docker
  to Podman
- Check whether the image is a manifest or not
- Check whether the image exists in the local Podman image store
- Check whether the image exists in the local Docker image store

I also find its handling of tags a little inflexible as it requires the
local source image to have the same tags as the target image (whereas
the Podman push command allows us to push a local image tag to multiple
tags in the target registry).

In fact, the only benefit the action really has over calling the Podman
CLI push command directly is that it takes care of exposing the pushed
image digest as an output.

I've therefore decided to refactor the Publish Container workflow to
call the Podman CLI push command directly instead of using
`redhat-actions/push-to-registry`.