Upload Playwright test results in CI workflow
The primary reason for retaining these files as an artifact is to make
it easier to diagnose test failures. We therefore want this step to run
even if the workflow has already failed due to test failures.

There are, of course, other steps before the Playwright tests that can
cause the workflow to fail. I've therefore also set `if-no-files-found`
to `ignore` so that we don't get a build warning when there are no
Playwright results to upload.