Add user audit entry model and table
In preparation for the introduction of more user administration
functions, I have decided to introduce a dedicated audit log for
operations performed on user accounts. This will include some operations
such as password changes and resets that are currently logged in the
security events table.

I've used a custom value converter for `UserOperationType` so that the
string values can be snake-cased like the event names in the security
events table.

I've kept the default cascade delete behaviour on the two foreign keys
to the users table as this makes it easier to clean up users in tests.
In production, the application will only allow users to be deactivated,
not deleted, so this shouldn't result in the deletion of audit data
outside of test environments.