Add dedicated filter type for IP address fields
Previously IP address filter fields were configured to use
`StringOperationFilterInputType`. However, that type defines some
operations (such as `contains`, `startsWith` and `endWith`) that cannot
be applied to fields with the `IPAddress` runtime type.

To address this I've added a dedicated
`IPAddressOperationFilterInputType` that only defines operations that
make sense for the `IPAddress`. The implementation of this class is
largely based on `IdOperationFilterInputType` [1].

[1] https://github.com/ChilliCream/graphql-platform/blob/65a4eb230b6a577a08c50e11febcb9f2b52f309f/src/HotChocolate/Data/src/Data/Filters/Types/IdOperationFilterInputType.cs#L6