Prevent deactivated users for resetting their password
Since a successful password reset results in the user being signed in,
we need to ensure that deactivated users are not able to complete the
password reset flow.

The earliest opportunity to do so is when sending the password reset
link. However, even if we did send a different email to deactivated
users that didn't include a password reset link, because the user could
become deactivated in the time between the link being sent and the user
completing the password reset flow, we would still need to add a final
check at the point of resetting the password.

Therefore, to keep things as simple as possible, I've opted to only
check whether the user is active when redeeming the password reset link.
This does make the unhappy path slightly longer than necessary in some
cases, but very few users are likely to be impacted in practice.

As part of this change I've introduced a `PasswordResetResult` class
that allows both `CanResetPassword` and `ResetPassword` in
`IPasswordAuthenticationService` to return the affected user on success
or a reason on failure, which means that the `ResetPassword` method no
longer needs to throw an exception on failure.