Remove `Build*Args` methods from `PasswordAuthenticationServiceTests`
All these methods did was call the model factory to generate a dummy
parameter values. Initialising those values directly in each test
introduces a small amount of code duplication, but that feels like a
price worth paying for the improved readability that comes from removing
that extra level of indirection.